# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Task Management backend server",
        "title": "heisenflow-System",
        "contact": {
            "name": "HeisenGo",
            "url": "https://github.com/HeisenGo"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new task for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Create task",
                "parameters": [
                    {
                        "description": "Task details",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.UserTask"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "response: details of created task",
                        "schema": {
                            "$ref": "#/definitions/presenter.CreateTaskResp"
                        }
                    },
                    "400": {
                        "description": "error: bad request, invalid task details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "error: forbidden, permission denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "error: bad gateway, not a member, user not found, board not found, or other error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/dependency": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a dependency between tasks for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Add task dependency",
                "parameters": [
                    {
                        "description": "Dependency details",
                        "name": "dependency",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.DependentTasks"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "response: details of added task dependency",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error: bad request, invalid dependency details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "error: forbidden, permission denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "error: bad gateway, circular dependency, task not found, or other error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/reorder": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reorder the tasks of a board for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Reorder Tasks",
                "parameters": [
                    {
                        "description": "Reorder Tasks Request",
                        "name": "ReorderColumnsRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.ReorderTasksReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tasks reordered successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/presenter.TaskReorderRespItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, invalid reorder details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden, permission denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/{taskID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the full details of a task by its ID for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get full task by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "taskID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task successfully fetched",
                        "schema": {
                            "$ref": "#/definitions/presenter.FullTaskResp"
                        }
                    },
                    "400": {
                        "description": "Bad request, invalid task ID or user not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/{taskID}/column": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the column of a task by its ID for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Update task column by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "taskID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Task Column Request",
                        "name": "UpdateTaskColReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.UpdateTaskColReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task successfully updated",
                        "schema": {
                            "$ref": "#/definitions/presenter.UpdatedTaskResp"
                        }
                    },
                    "400": {
                        "description": "Bad request, invalid task ID or column ID, or dependent task issues",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "presenter.CreateTaskResp": {
            "type": "object",
            "properties": {
                "assignee_user_id": {
                    "type": "string"
                },
                "board_id": {
                    "type": "string"
                },
                "column_id": {
                    "type": "string"
                },
                "dependsOn": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.DependTaskResp"
                    }
                },
                "description": {
                    "type": "string"
                },
                "end_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "parent_id": {
                    "description": "can be null for tasks not sub tasks",
                    "type": "string"
                },
                "start_at": {
                    "type": "string"
                },
                "story_at": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "presenter.DependTaskResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "presenter.DependentTasks": {
            "type": "object",
            "required": [
                "depends_on_task_ids",
                "task_id"
            ],
            "properties": {
                "depends_on_task_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "task_id": {
                    "type": "string"
                }
            }
        },
...
        "presenter.FullTaskResp": {
            "type": "object",
            "properties": {
                "dependencies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.TaskDependTaskResp"
                    }
                },
                "description": {
                    "type": "string"
                },
                "end_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "parent": {
                    "$ref": "#/definitions/presenter.TaskParentResp"
                },
                "start_at": {
                    "type": "string"
                },
                "story_point": {
                    "type": "integer"
                },
                "subtasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.TaskSubTaskResp"
                    }
                },
                "title": {
                    "type": "string"
                },
                "user": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/presenter.TaskUserResp"
                        }
                    ]
                }
            }
        },
...
        "presenter.ReorderTaskItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "presenter.ReorderTasksReq": {
            "type": "object",
            "properties": {
                "column_id": {
                    "type": "string"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ReorderTaskItem"
                    }
                }
            }
        },
        "presenter.TaskDependTaskResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "presenter.TaskParentResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "presenter.TaskReorderRespItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "presenter.TaskSubTaskResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "presenter.TaskUserResp": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
        "presenter.UpdateTaskColReq": {
            "type": "object",
            "required": [
                "column_id"
            ],
            "properties": {
                "column_id": {
                    "type": "string"
                }
            }
        },
        "presenter.UpdatedTaskResp": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "end_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "start_at": {
                    "type": "string"
                },
                "story_point": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "presenter.UserTask": {
            "type": "object",
            "required": [
                "assignee_user_id",
                "board_id",
                "title"
            ],
            "properties": {
                "assignee_user_id": {
                    "type": "string"
                },
                "board_id": {
                    "type": "string"
                },
                "depends_on_task_ids": {
                    "description": "for tasks that this task depends on",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "desc": {
                    "type": "string"
                },
                "end_at": {
                    "type": "string"
                },
                "parent_id": {
                    "description": "for tasks that depend on this task",
                    "type": "string"
                },
                "start_at": {
                    "type": "string"
                },
                "story_point": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Type 'Bearer' followed by a space and your JWT token.\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}